Multiple improvements, basehaxx 2.0
Only a small (tricky) rop is now stored in the savefile, over the secret base messages area. This rop loads another one from the SD card.
The second rop does all the PASLR and gspwn stuff, moreover it loads the code from the SD card too.
You now have to press START to open the embedded menu, otherwise the *hax payload is automatically loaded (from the SD card).
An option to uninstall the exploit has been added to the menu.